Fix iSCSI functional tests in Fedora 18
There are two fix.

1. modprobe iscsi_target_mod before running tests in Fedora 18

In Fedora 18, iscsi_target_mod is not loaded by default. If the module
is not loaded, the rtslib will fail to configure iSCSI targets for the
storage backend and the test will fail. This patch load the module
explicitly before running the tests.

2. Wait for some time for VDSM to refreshing the device list

If two iSCSI tests are arranged back to back, we should wait for VDSM to
refresh the iSCSI session info. Otherwise the _getIqnDevs method fails.

Change-Id: If0ffd21cbf0b68fa19e4f9a84e399bf2c75039bc
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/12594
Tested-by: oVirt Jenkins CI Server
Reviewed-by: Dan Kenigsberg <danken@redhat.com>