image: placeholder optimization on preallocation
When the destination of a copy or move command is an NFS domain
we should try to optimize the placeholders creation avoiding
unnecessary prezeroing (only relevant case: RAW PREALLOCATED).
In such case in the past we were using a TEMPORARY_VOLUME_SIZE
to create a placeholder with a temporary size of 10Mb.
This is interfering with live storage migration as the following
volumes in the chain are inheriting the temporary size inside the
qcow header (virtual disk size).
With this patch we are instead forcing the creation of the volume
as SPARSE as this won't have any side effect on the qcow header.

Change-Id: I32811bd45320ef02bfc593a71dfdafc0b0550c7d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=910445
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12692
Reviewed-by: Dan Kenigsberg <danken@redhat.com>