Fix checking bonding options for missing bonds.
Right up until now, we were checking the options for a bond A
that would have to be potentially created by checking for the
existance of /sys/class/net/A/bonding/<opt>. Obviously, if the
bond had not been created at that point, the option validation
would fail.

This patch makes the option validation be done on whichever bond
is present on the system and, if none are, it creates the bond
to check the options.

Change-Id: I023a5bb8a52719559bb9d4716f25e0cba8b3530b
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Bug-Url: https://bugzilla.redhat.com/918666
Reviewed-on: http://gerrit.ovirt.org/12701
Reviewed-by: Dan Kenigsberg <danken@redhat.com>