Relink template hard links to meta and lease files
In file based domains every image that is based on a
template contains a hard link to the template volumes.

Since an export domain may contain images without the
backing template, when copying the template to the export
domain we need to relink the derived images to it to be able
to later on collapse the chain on import
(without qemu-img crashing).
In addition, in case we're importing a template to override a
corrupted template we also need to relink.

Change-Id: Idce0f3f1812fdf45efeeeffcccb7dc22b3b0d0f0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=864073
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12837
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>