remoteFileHandler: update the stale comments of import statements
When new child process of remoteFileHandler is started in unit test,
there is no vdsm package installed, so we can not import modules under
vdsm. There is an "if __name__ != 'main'" to conditionally not import
modules from vdsm package in the child process.

After commit d8778ad, which is reviewed at http://gerrit.ovirt.org/9217 ,
'../' is added to PYTHONPATH of the child process in unit test. The
working dir of the child is ./tests, so "from vdsm import X" will get
the module at ./tests/../vdsm/X.py . It can safely import modules under
the vdsm package now. There is no need to protect the import statements.
Then old comment explaining this becomes stale.

This patch drops the old comment, while keeps the import statements in
the "if", because the imported modules are not used in child process.

Change-Id: Ia6bb8e614db7db4a9e56ac59850bef2568468cf9
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/12872
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>