Custom properties per device.
This is the first commit in a series of patches that
will implement custom properties per device.
The first commit deals with hot plugging and unplugging
devices and enabling custom properties for those hook
points.

Patch Set 1:
1) All methods related to nic and disk hot plug and unplug
in hooks.py and libvirtvm.py now pass the params dictionary
onwards.

2) Added unit test 'deviceCustomProperties' under hooksTests.py,
that verifies that a custom property sent to a hook is indeed
available in that hook.

Patch Set 2:
Updated the man page to reflect the fact that custom properties
are now exposed as environment variables to device hooks

Patch Set 3:
Broke down long line in man page.

Change-Id: Id1684c3bcf8838b43c7856f3794f1bffd8c4b28c
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13124
Reviewed-by: Dan Kenigsberg <danken@redhat.com>