hook-macspoof: fix packaging
The macspoof is an optional package. When we configure without hook
option, is not installed to rpmbuild/BUILDROOT. However in the
vdsm.spec.in, it is not an optional package, so rpmbuild fails to
locate the related file in this case.

This patch moves the hook-macspoof file section to the optional hooks
part in vdsm.spec.in .

Change-Id: I37ddc1d398b3a40d2af10f27dcf557d03d003f7a
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/13085
Reviewed-by: Dan Kenigsberg <danken@redhat.com>