Encode '=' when persisting the Task.
"For a reason unknown (to me), task.py does not use pickle to
persist its data to file. We cannot fix that without breaking
backward-compatibility."
Danken, circa 2012.

Caveat Emptor:
This patch expands the string space that can be encoded when
a task is dumped and not tries to solve all task serialization
issues.
Since the old code ignored fields that can't be encoded,
we continue to assume that doing this no harms the system.
Based on this assumption we are disregarding the case of a task
load in a host with this code when the dump contains real '_eq_'.

Bug-url: https://bugzilla.redhat.com/show_bug.cgi?id=915068

Change-Id: I735d28974d3953aafaf4b5e5f1a25363d22b50c3
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13128
Tested-by: Daniel Paikov <paikov@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>