Reduce calling to ethtool.
Currently we do separate ethtool device object retrieval for getaddr,
getnetmask and getipv6addrs. This patch unifies it into a single
getIpInfo while leaving the api for the separatemethods available.

Furthermore, it simplifies the get method by moving the device
specific methods to their own methods.

As a bonus, this has about 25% perf when creating a netinfo.NetInfo
instance.

Change-Id: Ic2c161b112865ef5cd4842fe70f26a3eeab1c3dd
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13166
Reviewed-by: Dan Kenigsberg <danken@redhat.com>