SetupNetworks: remove redundant logging.
Most of setupNetworks was wrapped in a try catch the sole purpose
of which was to log any error that could happen. This is not
necessary anymore due to the fact that supervdsm.py (caller of
setupNetworks) has its setupNetworks call under logDecorator
which achieves the original purpose in a more generic way.

Change-Id: Ifd1c3ef1bd936de56934b2a474b6b5fb829868c0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13170
Reviewed-by: Dan Kenigsberg <danken@redhat.com>