configNetworks: network info gathering to netinfo
These three methods are dealing with netinfo information and
use netinfo instances (speciall ifaceUsers). This patch proposes
moving them to netinfo.

The other two methods could be left in configNetwork, but I would
do it with a leading underscore.

This is the first of a series of patches that build gradually
towards network reloaded.

Change-Id: I34a07f0ea33407a566bcc9c82e5ba6ca0c0701a6
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13171
Reviewed-by: Dan Kenigsberg <danken@redhat.com>