NetReload: ConfigWriter, ifup/down to configurator
NetworkReloaded mandates that the configuration done by
configNetworks shall be done using configurators that do the
step of modifying the state of the networking stack.

This patch moves ConfigWriter ifup and ifdown to the
ifcfg configurator module (in netconfg) while keeping everything
working. Next in the series there shall be patches that wrap
the methods that were moved into actions done to the Ifcfg
Configurator object.

Change-Id: I8f28fa717eb501f61ed922cf2571181ff3b42532
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13182
Reviewed-by: Dan Kenigsberg <danken@redhat.com>