NetReload: Moved addNetwork logic to the configurator
This change starts moving the logic of configNetwork API endpoints
towards the configurators.

configureBridge, configureVlan, configureBond and configureNic
after this patch have to take care of the underlying devices
explicitly in the configurator modules. However, the underlying
device handling will be abstracted to the network models in a
future patch. Thus, the code of configureBond for any configurator
such as Ifcfg is gonna be something like:

self.configWriter.addBonding(params)
for slave in bond.slaves:
    slave.configure(params)

Change-Id: I877eddc07a17f0bdd376df7bab8b19ffe52f72ba
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13268
Reviewed-by: Dan Kenigsberg <danken@redhat.com>