image: update the volume permissions when needed
Volumes leaves created in 2.2 did not have the group writeable bit
set. In order to be able to start VMs from old NFS volumes we need
to check (and eventually fix) the group permissions during the volume
preparation.

The code to address this issue was already present but commit 5a0b2c9
reintroduced the problem substituting the prepareVolume calls with
prepareImage.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=911417
Change-Id: I516d8240840e3aa32bd8074420e331b8dc4e6da7
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13331
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>