Make sr-iov hook use interface instead of hostdev
According to libvirt documentation, for the sr-iov virtual function
"nics" to have permanent and unique MAC addresses, the passthrough
should be done using <interface type='hostdev'> instead of
<hostdev>.

This change modifies the hook to take advantage of this libvirt
definition for stable sr-iovs.

The chown step is necessary to attach the virtual function to the VM
Otherwise, it is not possible to detach the virtual function from the
OS and attach it to the VM due to dynamic_ownership=0 (which is
necessary for network storage because dynamic ownership, ATM, always
sets ownership to root:root on VM destroy/hotunplug).

Change-Id: Ib2f88a8400fe4d230c0877428afef62df8e397db
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13541
Reviewed-by: Dan Kenigsberg <danken@redhat.com>