stats: report the domain version
Given that some features are available only since a specific domain
version (e.g. live snapshots require at least V3) and the upgrade
is asynchronous for the non-master domains (see upgradeStoragePool)
VDSM should provide to the engine a way to determine if a domain
(other than the master) has been upgraded or not.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=951057
Change-Id: I88884356eefcd5040deccbf34eab271be8dbc522
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13620
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>