sriov: Make the hook safe for concurrency
The current model uses a cache that is not locked and could
be overwritten by a second VM creation effectively preventing
the return of the chown to root:root.

This patch solves that by creating with exclusivity a file for
each virtual function that is attached (thus avoiding locking)
but having concurrency safety due to the OS guarantee that the
not more than a single processes can perform an open when the
flags O_CREAT | O_EXCL are set.

Change-Id: I2c6ad7f2ee53911312700396000e0aca07a917e9
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13641
Reviewed-by: Dan Kenigsberg <danken@redhat.com>