Bond destroying
If there is deleted network with bond other than bond[0-4] and no
other network is using that bond, the bond is destroyed. bond[0-4]
must be kept for backward compatibility, but we don't want to accumulate
any other bond.

Because we need same set of bondings in load_needed_modules.py.in, set of
bondings was added to netinfo and imported from there. The path to
bonding masters now also comes from netinfo.BONDING_MASTERS.

Change-Id: I0cd27f1ff9a62e202c3513e54df5481d1b72ed7c
Bug-Url: https://bugzilla.redhat.com/928442
Signed-off-by: Petr Sebek <psebek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13752
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>