Adding comment before getting libvirt connection
If libvirt-sock does not exist, and vdsm is restarted - vdsm tried 10
seconds to communicate with libvirt. In the log we see a gap for 10
seconds after the "I am .." print and before the libvirt error exception.
This delay, without knowing the specific location of vdsm process, can hide
the actual reason for the delay. If the service stopped before the
exception we'll see only the "I am.. " print without knowing where vdsm
got stuck.

Change-Id: I8b6d8c53826fbb82f07a59f15edf672e855ef75f
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13786
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>