vdsm: change default migration bandwidth in conf
Libvirt changed the default behaviour in migration bandwith
definition. Prior to versions 0.10.x it used "0" to define
default 32MBps limit, now it uses "0" to define unlimited
speed.

This patch preserves the former way, when the default was
32MBps. It needed in enterprise environments, where mass
migration with unlimited bandwith can have a serious impact
on the network infrastructure.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=925981
Signed-off-by: Peter V. Saveliev <peet@redhat.com>
Change-Id: If9f3ac98d1868504c04d80eea5fc50d1d8cb1939
Reviewed-on: http://gerrit.ovirt.org/13781
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>