vdsm: Delay desktop lock until we are sure it was a disconnect
This is a hack to mitigate the issue of spice-gtk not respecting the
configured secure channels [1]. Spice-gtk is always connecting first to
a non-secure channel and the server tells the client then to connect
to a secure channel. However as a result of this we're getting events
of false positive disconnects and we need to ensure that we're really
having a disconnected client
This timer is supposed to delay the call to lock the desktop of the
guest. And only lock it, if it there was no new connect.
This is detected by the clientIp being set or not.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=879352

Change-Id: I6af5cf868dc9caaa07134b2b296d985f06cf15f2
Bug-Url: https://bugzilla.redhat.com/905930
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13798
Reviewed-by: Dan Kenigsberg <danken@redhat.com>