agent: XML Character filtering improvement
The filtering of invalid characters for XML documents must not be
performed on unicode strings but on the encoded utf-8 string.
If we do not do this we'll be missing out on filtering everything
what would destroy the XML documents.
If there has been found an invalid character in the UTF-8 string
it _may_ cause the message to be discarded due to UTF-8 decoding
errors. However we don't want invalid strings to come through any
way.

Updated the unit tests for reflecting the changes.

Change-Id: I21b970c8552b3f238c6bfe9271c5c5a37d40ea4d
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13809
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Vered Volansky <vvolansk@redhat.com>