vdsmd.init: replace libvirt logs filter
Currently VDSM adds to libvirtd.conf a filter to include debug
messages from libvirt daemon. However, the log size is increasing quickly
with useless information about libvirt objects when the system is
running out of problem.

This patch will improve the logs for libvirt sources and remove
OBJECT_* messages to reduce the size of libvirtd.log.

Thanks to Pavel Zhukov for suggestion.

Change-Id: I31218b3a6ac2f0866df03c6f84c9303d1a53f745
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=920614
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13642
Reviewed-by: Pavel Zhukov <pzhukov@redhat.com>
Tested-by: Pavel Zhukov <pzhukov@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>