vdsmd: do not swallow libvirt error
Recently I've seen multiple cases of vdsm not starting up with a vague
error "Failed to define network filters on libvirt", when the underlying
issue was that libvirt has crashed.

It's better to have a line or two of log noise, in order to see that
libvirt could not have been reached.

Change-Id: I6a91061f52ec4eff24dd1e10f525fd3fed9185ef
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/13909
Reviewed-by: Moti Asayag <masayag@redhat.com>
Reviewed-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Tested-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>