Initializing libvirt connectivity in external thread to clientIF
Getting libvirt connection as part of startup can take long under high
load. Until clientIF is ready to serve API requests, the host state is
non-responsive. To avoid that, we make the connection in external
thread, and until all recover process is ready the host reports the its
on initialization state.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=949192
Change-Id: Id3e849db7685f80995ee56989f0bed405602e550
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/14018
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>