guestIF: Additional object strings filtering
Filtering the plain UTF-8 data is not enough. The filter has
to be applied also on the object parsed from json to be really
effective.

Example why this is required:

o = json.loads('{ "x": "\\u0000" }')

Now the 'o' python object looks like this:
{ "x": "\x00" }

Therefore we need to filter even the parsed object

Change-Id: Id64438812a19b25b67b8e50c5c56b0d776c4463e
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/14126
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>