vdsm-tool: unify ad-hoc execCmd implementations
There are various ad-hoc execCmd implementations in the modules of
vdsm-tool. Before the parent of this patch, execCmd is in storage.misc,
we can not import storage.misc.execCmd without modifying PYTHONPATH. The
parent of this patch moved execCmd to vdsm.utils, so this patch makes
use of the 'official' execCmd implementation in vdsm.utils for all the
modules of vdsm-tool.

Change-Id: I8ca64ec5dbfebe5c5090ec8fac46785a538b79d3
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/13917
Reviewed-by: Dan Kenigsberg <danken@redhat.com>