Issuing a real pvs cmd when there are no PVs belonging to a VG.
Assumed the vgName is an existent reacheable VG, must be PVs
belonging to it.
The pvs dict may be staled if this is a FC setup and/or
connectStorageServer was not issued and therefore the storage view
was not refreshed.

This patch is an alternative to other fixes for the same bug.

Change-Id: I22cd4f099f2232f718d09f923630fe9828d11233
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=955993
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/14634
Tested-by: Daniel Paikov <paikov@gmail.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>