storage: Fix the case when VDSM host is not a gluster peer
Currently when using glusterSD, VDSM has the need to know the
transport type of gluster volume for which it uses gluster cli
which assumes that VDSM host is part of gluster cluster (peer).
This may not be true always and imposes a constraint that VDSM
host must be a gluster peer.

This patch fixes the above by doing 2 things...

Adds support for --remote-host gluster cli option to volumeInfo
API. This helps retrieve gluster volume info from a host that
is not a gluster peer (Eg. VDSM host)

Adds support to use the ability of gluster cli (--remote-host) to
retrieve the gluster volume info even when VDSM host is not a
gluster peer.

Change-Id: Ia5967f42555db27c104cd64449fccdc9b5fbc779
Signed-off-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/14666
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>