Fix getStorageDomainInfo() logic.
Domains with 'role' == MASTER_DOMAIN should always be attached
to a pool.
Detect 'stale' master domains.
Not hiding errors anymore.
Catching specific errors.

BC: If the domain is a Master but it is not connected to the actual
(host) pool, master_ver, host_id and lver are not returned.

Required for making repoStats pool independent.

Change-Id: I8b0b2ad3dca19cf203d937c1a9f6a12ab0f1095f
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/14671
Reviewed-by: Ayal Baron <abaron@redhat.com>