Adding [start|stop]MonitoringDomain().
startMonitoringDomain() is added for monitoring a storage domain
without being connected to the pool and geting and ID for such
SD lockspace.

stopMonitoringDomain() stops the monitoring and releases the ID.

Monitoring results are gathered with repoStats().

Caveat Emptor:
SDs added to the monitor using startMonitoringDomain() should not
be part of the pool! i.e. the pool domain set and the added SDs set
should be disjoint.

Making repoStats pool independent.

Change-Id: I983d49b0a42cc06428ec75b7795d23abaa6ab84c
Signed-off-by: Eduardo Warszawski <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19762
Tested-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>