Fix bashisms in vdsmd.init.in and ovirt_functions.sh
Fix vdsmd.init.in bashisms to make it run in dash, the default scripting
shell on Ubuntu. Use checkbashisms in Ubuntu devscripts package to check
it then fix one by one.

Some bashisms that are not going to be fixed in this patch:
1. "local" is not in POSIX but dash supports it, however old dash does
not support "local x y", so we have to "local x; local y".
2. gettext related bashisms, they do not harm in Ubuntu.

[1] https://launchpad.net/ubuntu/+source/devscripts
[2] https://wiki.ubuntu.com/DashAsBinSh#I_am_a_developer._How_can_I_avoid_this_problem_in_future.3F
[3] http://mywiki.wooledge.org/Bashism

Change-Id: I0eaa557b939bbd06f2cf661f9d95e4d907bbf605
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/14825
Reviewed-by: Dan Kenigsberg <danken@redhat.com>