netmodels: add a backward reference to NetDevice's master device
The netmodels keeps reference to slave device for a master device, but
the slave device doesn't have a backward reference. This patch adds
the reference on configuring, which can make it easier to get its usage
on removing an interface. In the coming delNetwork implementation based
on netmodels, we just call netdevice.remove() to delete a network entity.
So the NetDevice object should contain all the information needed by
the remove operation. We have different handling on removal when an
interface is attached to different types of masters. For example, if the
master of an inferface is Bond, Vlan, or Bridge, then we don't need
consider removing IP configuration on it because it could not be the top
object of the network hierarchy.

Change-Id: Iba8958549a301c9faac70a32cbafdf320128c85a
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/14870
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>