api: compat: fix api breakage of 'config' and 'netinfo' used by bootstrap
rhevm-3.0 depends on previous vdsm-bootstrap which uses config and
netinfo packages.

config and netinfo packages are now moved to the 'formal' vdsm api located
at vdsm namespace.

In order to allow rhevm-3.0 install vdsm-4.9.6 the legacy config and
netinfo should be available.

The old bootstrap code does not start the libvirtd service before
using the netinfo module.

The current vdsm.netinfo needs running libvirtd or it fails.

This patch starts the libvirtd service during the import of the
compatibility module.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=867922
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=870734
Change-Id: I57754ce86f1f1e5978ff7e665c88218ee20d7ba3
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Signed-off-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35721
Reviewed-by: Dan Kenigsberg <danken@redhat.com>