Extend vdsm-tool commands - retrieving host uuid by utils.getHostUUID
Deploying an host requires to verify if an host has vdsm installed.
The command 'vdsm-tool vdsm-id' returns vdsm uuid if vdsm is installed.
It doesn't require running vdsm instance.
This patch also adds 'legacy' parameter to getHostUUID for backward
compatibility, if set to false the mac address won't be added to the
uuid.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=964967
Change-Id: Ia8cab7f55697e07479a4f35c161bd836925374e8
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/14866
Reviewed-by: Dan Kenigsberg <danken@redhat.com>