vdsm-tool: service probing do not trigger abrt warning on RHEL6
When an service management operation fails, the service module raises a
RuntimeError, this triggers abrt warning of un-caught Python exceptions.
This is good for reporting error when fail to start a service, but
probing command such as "service-status" should not raise un-caught
exception if the service is down. It should report the error code and
leave the caller to decide what to do (probably start that service).

In this patch we define a specific exception for service operations,
then for service probing operation, it swallows the exception, prints
error message and returns 1. For other management operations, the
exception is still raised.

Change-Id: Iddac6c6d1b6f47fe653ca4a86cc3a89f5ed4efb6
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/15098
Reviewed-by: Dan Kenigsberg <danken@redhat.com>