vdsm.tool.service: fix _systemctlNative
_systemctlNative lists systemd units and check if the given service is
managed by systemd. There is a bug in the checking code. When
splitting words, it does not fetch the word from the resulting array.
The first word is the name of the unit/service. This patch fix the
problem.

Another small problem is it does not format the error message correctly,
this patch fix the formatting.

At last it fixes a bad doc string.

Change-Id: I3a3b903786d431ba306bb2cf3f7f35d6de38ed42
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/15305
Reviewed-by: Dan Kenigsberg <danken@redhat.com>