Making getAllTasks[List and Statuses] an SPM command
Currently VDSM does not manage HSM tasks. This verb is used only to poll
SPM tasks. If the host is not SPM, an exception will be raised. Backwards
compatibility is not required as the verb API is not changed and the
behaviour stays as expected in older clusters (Instead of returning empty
list for non-spm hosts, vdsm will raise exception as expected).

Change-Id: I46a107d8c7ddcd0b4e74045a27ef3683f6a78489
Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=880547
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/15345
Reviewed-by: Ayal Baron <abaron@redhat.com>