vdsm: vmChannels: Do not use epoll unregister
We don't use epoll.unregister for the deregistration from epoll anymore.
According to the epoll documentation, epoll will remove the fd when the
fd was closed.
Since we're always closing the sockets where this is used and we
otherwise might run into race conditions we're no longer calling
epoll.unregister to avoid attempts to unregister fds which might no
longer belong to us.

The problem on this race condition is that we're handling file descriptors
by numbers. When we're ending the GuestAgent by calling stop() it will send
the fd to the listener thread by inserting it into a queue so the listener
can handle it the next time. However we're immediately closing the socket and
the OS is free to reuse the fd. By the time we're going to unregister the fd
from epoll the fd might have been already reused for our process by the OS
because it reassigns the lowest available fileno.

To solve this issue we're using the capability of epoll to detect closed
fds and rely on it's mechanism to unregister them. By that we're not running
into the issue of unregistering the fd from fd which might have been already
already assigned to a different object.

Change-Id: Iacb8dee05e179405d7946bef7329ca406223744a
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/15393
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Lee Yarwood <lyarwood@redhat.com>