netinfo: Simplify the check if an interface still has any other users.
In current oVirt network models, a nic device has four kinds
of usages:
    * bonding's slave
	* vlan's slave
	* bridge's port
	* bridgeless network's interface
A bond device has the same usages as nic except the first one.

And an interface(nic or bonding) could be multiplexed by different usages.
But bonding and bridge need use the interface exclusively. So only two
combinations are valid:
	* multiple vlan users
	* multiple vlan users with one bridgeless network user

When we remove one user of an interface, we need check if the interface
still has by other users except the user being removed currently. According
to that, we decide if restore the interface to initial state or destroy it
(for bonding).

The original code can work fine both with the netinfo instantiated before
or after the user is removed. In the new code, it needs an updated netinfo
instance, which doesn't include the user is already removed. And it doesn't
require any information about the user removed.

In the original implementation, the checks for nic and bonding are handled differently
because it also checks the bond users if a nic is a slave of bond. Actually,
it's not neccessary either. If the bond has other users, we will not remove its
slave. That means the check will not be invoked. So for the check of nic, it
doesn't need care about the usage of its bond master, because it's done in bond
level.

Change-Id: Ie70cbfcc4c561c98f2e90685329900796517f933
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/15416
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Tested-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>