Multiple Gateways[2/2]: configNetwork integration
This patch provides the integration of sourceRoute with the
configurator flow.

For static routes, multiple gateways uses the ifcfg configurator
and its configWriter to write:
    /etc/sysconfig/network-scripts/{route,rule}-netname
Those files are taken care of by ifup/ifdown.

For dynamic routes (DHCP) it has a hook in
/etc/dhcp/dhclient.d/sourceRoute.sh that is automatically executed
by dhcpd when it obtains a lease for a device. The hook will call
directly the sourceRoute.py command line UI that will delegate
the actions to the iproute2 configurator. For this hook to work,
selinux needs a few policy updates that have been requested in:

https://bugzilla.redhat.com/show_bug.cgi?id=977856

The selinux-policy changes are needed due to the fact that the python
code in sourceRoute.py is being executed from the dhcp_t context
and not from the usual service, reason for which currently there was
no policy to allow such context to access several parts of the system
via python.

The integration is done at the configureDevice level because it is
the only sensible choice considering that the static file removal
must be done between ifdown and ifup (so that ifdown takes down the
rules and ifup doesn't have them anymore so it does not apply them)
and the fact that configureDevice for nics and bonds must leave the
device up, i.e., they do ifdown and ifup inside the method.

Patch Set 2:
- delNetwork now checks that the bootproto is not dhcp
- In sourceRoute.py: Removed check if the interface has
  a libvirt network on top of it because it was preventing
  DHCP interfaces from going up properly. Will need to think
  of a different fix.

Patch Set 4:
- Fixed the Makefile and spec for dhcp hook installation.

Patch Set 5:
- Fixed the network deletion. It was trying to remove routes for an
  already unexistant network device.

Patch Set 7:
- Fixed dhcp flow by changing when we do libvirt network addition
  and deletion.

Patch Set 8:
- Addressed review comments. Fixed the ifup ifdown flow when using
  the ifcfg configurator (and the iproute2 for dhcp).

Patch Set 9:
- Addressed review comments and squashed:
      http://gerrit.ovirt.org/#/c/16087/1
      I09aec8fa9a2ef2d61609a0a807f4e7130978fbfb
  into this patch.

Change-Id: Id4415c6ec91ac49c56703d59b99d3a21b328015b
Signed-off-by: Assaf Muller <amuller@redhat.com>
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/15528
Reviewed-by: Dan Kenigsberg <danken@redhat.com>