vm: add the live diskSizeExtend method
This patch adds the diskSizeExtend HSM method to implement the live disk
extension.

 returnCode = diskSizeExtend(vmUUID, driveSpecs, newSize)

 driveSpecs =  {
   'poolID':   '<spUUID>',
   'domainID': '<sdUUID>',
   'imageID':  '<imgUUID>',
   'volumeID': '<volUUID>',
 }

 returnCode = {
   'status': {'code': 0, 'message': 'Done'},
   'size': '<newSize>',
 }

The driveSpecs (usually poolID, domainID, imageID, volumeID) indentify
the disk to extend. The driveSpecs as GUID is not supported.
The returnCode also returns the newSize of the disk seen by the guest
(might be different from the requested one, e.g. rounded by QEMU).

The returnCode might also return the following specific errors:

- noVM(1): Virtual machine does not exist
- resizeErr(58): Wrong resize disk parameter
- updateDevice(56): Failed to update device

Change-Id: I326f0e71d53382a49eb3b43cdf0bc0472f71abdc
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/15614
Reviewed-by: Ayal Baron <abaron@redhat.com>